requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-client-cache==0.11.1
aiosignal==1.3.1
aiosqlite==0.20.0
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
backports.tarfile==1.2.0
bandit==1.7.10
beautifulsoup4==4.12.3
bitarray==3.0.0
black==24.10.0
bleach==6.1.0
blinker==1.8.2
boolean.py==4.0
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.3.3
catboost==1.2.5
cattrs==24.1.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
-e git+https://github.com/BrandynHamilton/Chart-Builder.git@9890f0e59af91f0f52d51aee7cb96c377a015d35#egg=chart_builder
ckzg==1.0.2
clarabel==0.9.0
cleo==2.1.0
click==8.1.7
cloudpickle==3.0.0
cmdstanpy==1.2.3
colorama==0.4.6
colorcet==3.1.0
comm==0.2.2
contourpy==1.2.1
convertdate==2.4.0
crashtest==0.4.1
croniter==1.4.1
crypto_cpp_py==1.4.4
cvxpy==1.5.1
cycler==0.12.1
cyclonedx-python-lib==7.6.2
Cython==3.0.10
cytoolz==0.12.3
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
deepdiff==8.1.1
defusedxml==0.8.0rc2
Deprecated==1.2.14
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
docutils==0.21.2
dulwich==0.21.7
dune_client==1.7.5
ecdsa==0.18.0
ecos==2.0.13
ephem==4.1.5
et_xmlfile==2.0.0
eth-account==0.11.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==1.0.1
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.1.0
executing==2.0.1
Farama-Notifications==0.0.4
fastapi==0.115.6
fastjsonschema==2.20.0
filelock==3.15.1
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
flipside==2.0.8
fonttools==4.53.0
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-analytics-data==0.18.15
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.65.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
gymnasium==0.29.1
h11==0.14.0
h5py==3.11.0
hexbytes==0.3.1
holidays==0.50
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflection==0.5.1
inotify==0.2.10
installer==0.7.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_scheduler==2.10.0
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kaleido==0.2.1
keras==3.3.3
keyring==24.3.1
kiwisolver==1.4.5
lark==1.1.9
libclang==18.1.1
license-expression==30.4.0
lightgbm==4.4.0
lru-dict==1.2.0
LunarCalendar==0.0.9
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.0.1
marshmallow_dataclass==8.6.1
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
ml-dtypes==0.3.2
-e git+https://github.com/BrandynHamilton/Data-Pipeline.git@aed850814a23d19e40bd372144c43f4344a6100b#egg=modular_pipeline
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
ndjson==0.3.1
nest-asyncio==1.6.0
networkx==3.3
nh3==0.2.19
nose==1.3.7
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
openai==1.59.6
openbb==4.3.5
openbb-benzinga==1.3.5
openbb-bls==1.0.3
openbb-cftc==1.0.3
openbb-commodity==1.2.6
openbb-core==1.3.8
openbb-crypto==1.3.5
openbb-currency==1.3.5
openbb-derivatives==1.3.5
openbb-econdb==1.2.5
openbb-economy==1.3.5
openbb-equity==1.3.5
openbb-etf==1.3.5
openbb-federal-reserve==1.3.5
openbb-fixedincome==1.3.5
openbb-fmp==1.3.5
openbb-fred==1.3.5
openbb-imf==1.0.2
openbb-index==1.3.5
openbb-intrinio==1.3.5
openbb-news==1.3.5
openbb-oecd==1.3.5
openbb-platform-api==1.0.4
openbb-polygon==1.3.5
openbb-regulators==1.3.5
openbb-sec==1.3.5
openbb-tiingo==1.3.5
openbb-tradingeconomics==1.3.5
openbb-us-eia==1.0.0
openbb-yfinance==1.3.6
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.11.0
orderly-set==5.2.3
osqp==0.6.7
overrides==7.7.0
packageurl-python==0.16.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parsimonious==0.10.0
parso==0.8.4
pathspec==0.12.1
patsy==0.5.6
pbr==6.1.0
peewee==3.17.5
pexpect==4.9.0
pillow==10.3.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.7.3
pkginfo==1.12.0
platformdirs==4.2.2
plotly==5.22.0
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
poseidon_py==0.1.4
posthog==3.7.5
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.2.1
prophet==1.1.5
proto-plus==1.25.0
protobuf==5.28.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-serializable==1.1.2
pyairtable==2.3.5
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.26.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
PyMeeus==0.5.12
pyngrok==7.1.6
pyparsing==3.1.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.18
pytz==2024.1
pyunormalize==16.0.0
pywin32==306
pywin32-ctypes==0.2.3
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qdldl==0.1.7.post2
qtconsole==5.5.2
QtPy==2.4.1
RapidFuzz==3.11.0
readme_renderer==44.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-cache==1.2.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.1
rlp==4.0.1
rpds-py==0.18.1
rsa==4.9
ruff==0.7.4
scikit-learn==1.5.0
scipy==1.13.1
scs==3.2.4.post2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools-git==1.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.36
stable_baselines3==2.3.2
stack-data==0.6.3
stanio==0.5.0
starknet-py==0.22.0
starlette==0.41.3
statsmodels==0.14.4
stevedore==5.3.0
streamlit==1.35.0
svgwrite==1.4.3
sympy==1.11.1
tbb==2021.12.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
tomlkit==0.13.2
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
trove-classifiers==2025.1.10.15
twine==6.0.1
types-Deprecated==1.2.9.20240311
types-python-dateutil==2.9.0.20240316
types-PyYAML==6.0.12.20240808
types-requests==2.31.0.6
types-setuptools==73.0.0.20240822
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uniswap-python==0.7.1
uri-template==1.3.0
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.19
uuid7==0.1.0
uvicorn==0.32.1
virtualenv==20.28.1
watchdog==4.0.1
wcwidth==0.2.13
web3==6.20.3
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
xgboost==2.1.0
xlrd==2.0.1
xmltodict==0.13.0
yarl==1.18.3
yfinance==0.2.51
zipp==3.19.2
